remove renormalization of features from adhocprovider
