+++ libraries/appframeworks/portals/src/lib/utils/adhoc-feature-utils.ts
@@ -1,4 +1,3 @@
-import md5 from "md5";
 import type { Feature, FeatureCollection } from "geojson";
 import type { GeoJSONSourceSpecification } from "maplibre-gl";
 
@@ -15,178 +14,12 @@ const isGeoJsonSource = (
   source !== null &&
   (source as { type?: unknown }).type === "geojson";
 
-const hasGeoJsonFeatureId = (feature: Feature): boolean => {
-  const id = feature.id;
-  return typeof id === "string" || typeof id === "number";
-};
-
-const createGeneratedGeoJsonFeatureId = (
-  feature: Feature,
-  featureIndex: number
-): string => {
-  try {
-    return `autogen:${md5(JSON.stringify(feature))}`;
-  } catch {
-    return `autogen:${md5(String(featureIndex))}`;
-  }
-};
-
-const ensureGeoJsonFeatureIds = (
-  geojson: Feature | FeatureCollection
-): {
-  geojson: Feature | FeatureCollection;
-  changed: boolean;
-  generatedFeatureIds: string[];
-} => {
-  if (geojson.type === "FeatureCollection") {
-    let changed = false;
-    const generatedFeatureIds: string[] = [];
-    const usedIds = new Set<string>(
-      geojson.features.flatMap((feature) =>
-        hasGeoJsonFeatureId(feature) ? [String(feature.id)] : []
-      )
-    );
-    const nextFeatures = geojson.features.map((feature, featureIndex) => {
-      if (!hasGeoJsonFeatureId(feature)) {
-        changed = true;
-        const baseGeneratedFeatureId = createGeneratedGeoJsonFeatureId(
-          feature,
-          featureIndex
-        );
-        let generatedFeatureId = baseGeneratedFeatureId;
-        let duplicateSuffix = 1;
-        while (usedIds.has(generatedFeatureId)) {
-          generatedFeatureId = `${baseGeneratedFeatureId}:${duplicateSuffix}`;
-          duplicateSuffix += 1;
-        }
-        usedIds.add(generatedFeatureId);
-        generatedFeatureIds.push(generatedFeatureId);
-        return {
-          ...feature,
-          id: generatedFeatureId,
-        };
-      }
-      usedIds.add(String(feature.id));
-      return feature;
-    });
-    if (!changed) {
-      return { geojson, changed: false, generatedFeatureIds: [] };
-    }
-    return {
-      geojson: {
-        ...geojson,
-        features: nextFeatures,
-      },
-      changed: true,
-      generatedFeatureIds,
-    };
-  }
-
-  if (hasGeoJsonFeatureId(geojson)) {
-    return { geojson, changed: false, generatedFeatureIds: [] };
-  }
-
-  const generatedFeatureId = createGeneratedGeoJsonFeatureId(geojson, 0);
-  return {
-    geojson: {
-      ...geojson,
-      id: generatedFeatureId,
-    },
-    changed: true,
-    generatedFeatureIds: [generatedFeatureId],
-  };
-};
-
 export const isAdhocMapLibreStyleFeature = (
   feature: AdhocFeature
 ): feature is AdhocFeature & {
   kind: "maplibre-style";
   data: CarmaMapLibreStyleData;
 } => feature.kind === "maplibre-style";
-
-export const normalizeAdhocFeatureGeoJsonFeatureIds = (
-  feature: AdhocFeature
-): {
-  feature: AdhocFeature;
-  generatedGeoJsonFeatureIds: string[];
-} => {
-  if (!isAdhocMapLibreStyleFeature(feature)) {
-    return {
-      feature,
-      generatedGeoJsonFeatureIds: [],
-    };
-  }
-
-  const sources = feature.data.sources;
-  if (!sources) {
-    return {
-      feature,
-      generatedGeoJsonFeatureIds: [],
-    };
-  }
-
-  let hasChanges = false;
-  const generatedGeoJsonFeatureIds: string[] = [];
-  const normalizedSources = Object.fromEntries(
-    Object.entries(sources).map(([sourceId, sourceValue]) => {
-      if (
-        !isGeoJsonSource(sourceValue) ||
-        typeof sourceValue.data !== "object" ||
-        sourceValue.data === null
-      ) {
-        return [sourceId, sourceValue];
-      }
-
-      const sourceData = sourceValue.data as
-        | Feature
-        | FeatureCollection
-        | Record<string, unknown>;
-      if (
-        sourceData.type !== "Feature" &&
-        sourceData.type !== "FeatureCollection"
-      ) {
-        return [sourceId, sourceValue];
-      }
-
-      const normalized = ensureGeoJsonFeatureIds(
-        sourceData as Feature | FeatureCollection
-      );
-      if (normalized.generatedFeatureIds.length > 0) {
-        generatedGeoJsonFeatureIds.push(...normalized.generatedFeatureIds);
-      }
-      if (!normalized.changed) {
-        return [sourceId, sourceValue];
-      }
-
-      hasChanges = true;
-      return [
-        sourceId,
-        {
-          ...sourceValue,
-          data: normalized.geojson,
-        },
-      ];
-    })
-  );
-
-  if (!hasChanges) {
-    return {
-      feature,
-      generatedGeoJsonFeatureIds,
-    };
-  }
-
-  return {
-    feature: {
-      ...feature,
-      data: {
-        ...feature.data,
-        sources: normalizedSources as CarmaMapLibreStyleData["sources"],
-      },
-    },
-    generatedGeoJsonFeatureIds,
-  };
-};
 
 export const getMapLibreLayerInfo = (feature: AdhocFeature) => {
   if (!isAdhocMapLibreStyleFeature(feature)) return undefined;
